added a todo note, added bones for message function in server
We're supposed to reply that the user is not joined to the channel if
that is the case when it tries to send a message to a channel.

We could do it by adding the Pid or nickname to the msg_from_GUI tuple,
or by creating another function that communicates the server with the
client, specifically to tell if the client is in a channel or not.